toolkit/components/osfile/tests/mochi/test_osfile_front.xul

<?xml version="1.0"?>
<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<window title="Testing OS.File on a chrome worker thread"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="test();">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="application/javascript"
          src="worker_handler.js"/>
  <script type="application/javascript">
  <![CDATA[

let worker;
let main = this;

function test() {
  info("test_osfile_front.xul: Starting test");

  // Test the OS.File worker

  worker = new ChromeWorker("worker_test_osfile_front.js");
  SimpleTest.waitForExplicitFinish();
  info("test_osfile_front.xul: Chrome worker created");
  dump("MAIN: go\n");
  worker_handler(worker);
  worker.postMessage(0);
  ok(true, "test_osfile_front.xul: Test in progress");
};
]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display:none;"></div>
    <pre id="test"></pre>
  </body>
  <label id="test-result"/>
</window>
